Moving route handling to routes/
