+++ package.json
@@ -28,6 +28,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "faker": "^4.1.0",
+    "morgan": "^1.10.0",
     "mysql": "^2.18.1",
     "neat-csv": "^5.2.0",
     "prop-types": "^15.7.2",
